new_transformer/requirements.txt
beautifulsoup4==4.11.1
boto3==1.21.46
botocore==1.24.46
certifi==2021.10.8
charset-normalizer==2.0.12
Cython==3.0.0a10
future==0.18.2
geographiclib==1.52
geopy==2.2.0
idna==3.3
jmespath==1.0.0
lxml==4.8.0
numpy==1.22.3
pandas==1.4.2
pbr==5.8.1
psycopg2==2.9.3
py4j==0.10.9.3
pymongo==4.1.1
pyspark==3.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
scipy==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
